Module google_gamesconfiguration1_configuration::api

Structs

AchievementConfiguration
An achievement configuration resource.

AchievementConfigurationDeleteCall
Delete the achievement configuration with the given ID.

AchievementConfigurationDetail
An achievement configuration detail.

AchievementConfigurationGetCall
Retrieves the metadata of the achievement configuration with the given ID.

AchievementConfigurationInsertCall
Insert a new achievement configuration in this application.

AchievementConfigurationListCall
Returns a list of the achievement configurations in this application.

AchievementConfigurationListResponse
A ListConfigurations response.

AchievementConfigurationMethods
A builder providing access to all methods supported on achievementConfiguration resources. It is not used directly, but through the GamesConfiguration hub.

AchievementConfigurationUpdateCall
Update the metadata of the achievement configuration with the given ID.

GamesConfiguration
Central instance to access all GamesConfiguration related resource activities

GamesNumberAffixConfiguration
A number affix resource.

GamesNumberFormatConfiguration
A number format resource.

ImageConfiguration
An image configuration resource.

ImageConfigurationMethods
A builder providing access to all methods supported on imageConfiguration resources. It is not used directly, but through the GamesConfiguration hub.

ImageConfigurationUploadCall
Uploads an image for a resource with the given ID and image type.

LeaderboardConfiguration
An leaderboard configuration resource.

LeaderboardConfigurationDeleteCall
Delete the leaderboard configuration with the given ID.

LeaderboardConfigurationDetail
A leaderboard configuration detail.

LeaderboardConfigurationGetCall
Retrieves the metadata of the leaderboard configuration with the given ID.

LeaderboardConfigurationInsertCall
Insert a new leaderboard configuration in this application.

LeaderboardConfigurationListCall
Returns a list of the leaderboard configurations in this application.

LeaderboardConfigurationListResponse
A ListConfigurations response.

LeaderboardConfigurationMethods
A builder providing access to all methods supported on leaderboardConfiguration resources. It is not used directly, but through the GamesConfiguration hub.

LeaderboardConfigurationUpdateCall
Update the metadata of the leaderboard configuration with the given ID.

LocalizedString
A localized string resource.

LocalizedStringBundle
A localized string bundle resource.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.