Struct google_gamesconfiguration1_configuration::AchievementConfiguration

pub struct AchievementConfiguration {
    pub achievement_type: Option<String>,
    pub steps_to_unlock: Option<i32>,
    pub kind: Option<String>,
    pub initial_state: Option<String>,
    pub token: Option<String>,
    pub draft: Option<AchievementConfigurationDetail>,
    pub published: Option<AchievementConfigurationDetail>,
    pub id: Option<String>,
}

This is a JSON template for an achievement configuration resource.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list achievement configurations (none)

update achievement configurations (request|response)

insert achievement configurations (request|response)

get achievement configurations (response)

patch achievement configurations (request|response)

delete achievement configurations (none)

Fields

achievement_type

The type of the achievement. Possible values are:
- "STANDARD" - Achievement is either locked or unlocked. - "INCREMENTAL" - Achievement is incremental.

steps_to_unlock

Steps to unlock. Only applicable to incremental achievements.

kind

Uniquely identifies the type of this resource. Value is always the fixed string gamesConfiguration#achievementConfiguration.

initial_state

The initial state of the achievement. Possible values are:
- "HIDDEN" - Achievement is hidden. - "REVEALED" - Achievement is revealed. - "UNLOCKED" - Achievement is unlocked.

token

The token for this resource.

draft

The draft data of the achievement.

published

The read-only published data of the achievement.

id

The ID of the achievement.