Struct google-gamesconfiguration1_configuration::GamesNumberFormatConfiguration

pub struct GamesNumberFormatConfiguration {
    pub currency_code: String,
    pub number_format_type: String,
    pub num_decimal_places: i32,
    pub suffix: GamesNumberAffixConfiguration,
}

This is a JSON template for a number format resource.

This type is not used in any activity, and only used as part of another schema.

Fields

currency_code

The curreny code string. Only used for CURRENCY format type.

number_format_type

The formatting for the number. Possible values are:
- "NUMERIC" - Numbers are formatted to have no digits or a fixed number of digits after the decimal point according to locale. An optional custom unit can be added. - "TIME_DURATION" - Numbers are formatted to hours, minutes and seconds. - "CURRENCY" - Numbers are formatted to currency according to locale.

num_decimal_places

The number of decimal places for number. Only used for NUMERIC format type.

suffix

An optional suffix for the NUMERIC format type. These strings follow the same plural rules as all Android string resources.