Struct google_gamesconfiguration1_configuration::ImageConfiguration

pub struct ImageConfiguration {
    pub url: Option<String>,
    pub resource_id: Option<String>,
    pub kind: Option<String>,
    pub image_type: Option<String>,
}

This is a JSON template for an image configuration resource.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

upload image configurations (response)

Fields

url

The url for this image.

resource_id

The resource ID of resource which the image belongs to.

kind

Uniquely identifies the type of this resource. Value is always the fixed string gamesConfiguration#imageConfiguration.

image_type

The image type for the image.